refactor: move compute_pass_metrics into orchestrator/utils.py, drop eval_utils.py
compute_pass_metrics was the only thing in eval_utils.py; fold it into the existing
orchestrator/utils.py and delete the one-function module.
